requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
colorlog==6.9.0
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
logtail-python==0.3.2
msgpack==1.1.0
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyMySQL==1.1.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
